requirements.txt
praat-parselmouth
scipy
librosa
torch
numpy
tqdm
